Fix a memory leak in Connect.
This change adds a finalizer notifier for all GClosures created for
Connect.  When the GClosure is finalized, the closure and closure
context is removed from the internal closure map.  Removing the
contexts from this map removes the reflected func and user data from
Go's scope, allowing them to be garbage collected.