More performant {Idle,Timeout}Add
This commit changes the IdleAdd and TimeoutAdd functions to use the
g_full functions instead of the GClosure and g_source_attach API.

The reason for this change was because both of these removed APIs hog a
lot of CPU time for weird reasons and will eventually lag behind after a
long time of usage. The new functions don't have this issue at all.

A type check cache is also implemented just for IdleAdd and TimeoutAdd
functions, because these functions are potentially uesd for critical
functionality and should therefore be fast if possible.

glib.MainContext's Acquire method is also changed to return a boolean in
accordance to what's put in the GLib Main Loop documentation.