Fix dangerous memory reference again.
The fix made in e5dd2587cdc1 and outlined in bug #727 seems to have been
accidentally re-introduced in commit cf79ce840c3f (presumably because of
a merge conflict).  The same reproducer in the bug report can be used to
trigger this issue and validate the fix in this commit.